refactor: Cookies 삭제
