+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequest.java
@@ -4,7 +4,7 @@ import java.util.Arrays;
 import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
-import org.apache.coyote.http11.session.Cookies;
+import org.apache.coyote.http11.response.Cookie;
 
 public class HttpRequest {
 
@@ -44,8 +44,8 @@ public class HttpRequest {
         Map<String, String> cookies = Arrays.stream(cookieValue.split(COOKIE_CONNECTOR))
                 .collect(Collectors.toMap(cookie -> cookie.split(COOKIE_PARAMETER_DELIMITER)[KEY_INDEX],
                         cookie -> cookie.split(COOKIE_PARAMETER_DELIMITER)[VALUE_INDEX]));
-        if (cookies.containsKey(Cookies.JSESSIONID)) {
-            return Optional.of(cookies.get(Cookies.JSESSIONID));
+        if (cookies.containsKey(Cookie.JSESSIONID)) {
+            return Optional.of(cookies.get(Cookie.JSESSIONID));
         }
         return Optional.empty();
     }
